tests/wpt/mozilla/tests/mozilla/interfaces.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Interfaces exposed on the window</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// This is a list of all interfaces that are exposed to every webpage.
// Please only add things to this list with great care and proper review
// from the associated module peers.

// IMPORTANT: Do not change this list without review from
//            a JavaScript Engine peer!
var ecmaGlobals = [
  "Array",
  "ArrayBuffer",
  "Boolean",
  "DataView",
  "Date",
  "Error",
  "EvalError",
  "Float32Array",
  "Float64Array",
  "Function",
  "Infinity",
  "Int16Array",
  "Int32Array",
  "Int8Array",
  "InternalError",
  "Iterator",
  "JSON",
  "Map",
  "Math",
  "NaN",
  "Number",
  "Object",
  "Proxy",
  "RangeError",
  "ReferenceError",
  "RegExp",
  "Set",
  "StopIteration",
  "String",
  "SyntaxError",
  "TypeError",
  "Uint16Array",
  "Uint32Array",
  "Uint8Array",
  "Uint8ClampedArray",
  "URIError",
  "WeakMap",
];

// This is a list of all the DOM interfaces exposed to the window, with their inheritance:
// To add an interface to this list add the name as a key, and the parent name or null as the value.
// IMPORTANT: Do not change the list below without review from a DOM peer!
var interfaceNamesInGlobalScope = {
  "Attr": null,
  "Blob": null,
  "CanvasGradient": null,
  "CanvasRenderingContext2D": null,
  "CanvasPattern": null,
  "CharacterData": null,
  "CloseEvent": "Event",
  "CSS": null,
  "CSSStyleDeclaration": null,
  "DOMRect": null,
  "Comment": null,
  "Console": null,
  "CustomEvent": "Event",
  "DedicatedWorkerGlobalScope": null, // #2823
  "Document": null,
  "DocumentFragment": null,
  "DocumentType": null,
  "DOMException": null,
  "DOMImplementation": null,
  "DOMParser": null,
  "DOMTokenList": null,
  "DOMStringMap": null,
  "Element": null,
  "ErrorEvent": "Event",
  "Event": null,
  "EventTarget": null,
  "File": null,
  "FormData": null,
  "HTMLAnchorElement": "HTMLElement",
  "HTMLAppletElement": "HTMLElement",
  "HTMLAreaElement": "HTMLElement",
  "HTMLAudioElement": "HTMLElement",
  "HTMLBaseElement": "HTMLElement",
  "HTMLBodyElement": "HTMLElement",
  "HTMLBRElement": "HTMLElement",
  "HTMLButtonElement": "HTMLElement",
  "HTMLCanvasElement": "HTMLElement",
  "HTMLCollection": null,
  "HTMLDataElement": "HTMLElement",
  "HTMLDataListElement": "HTMLElement",
  "HTMLDialogElement": "HTMLElement",
  "HTMLDirectoryElement": "HTMLElement",
  "HTMLDivElement": "HTMLElement",
  "HTMLDListElement": "HTMLElement",
  "HTMLElement": null,
  "HTMLEmbedElement": "HTMLElement",
  "HTMLFieldSetElement": "HTMLElement",
  "HTMLFontElement": "HTMLElement",
  "HTMLFormElement": "HTMLElement",
  "HTMLFrameElement": "HTMLElement",
  "HTMLFrameSetElement": "HTMLElement",
  "HTMLHeadElement": "HTMLElement",
  "HTMLHeadingElement": "HTMLElement",
  "HTMLHRElement": "HTMLElement",
  "HTMLHtmlElement": "HTMLElement",
  "HTMLIFrameElement": "HTMLElement",
  "HTMLImageElement": "HTMLElement",
  "HTMLInputElement": "HTMLElement",
  "HTMLLabelElement": "HTMLElement",
  "HTMLLegendElement": "HTMLElement",
  "HTMLLIElement": "HTMLElement",
  "HTMLLinkElement": "HTMLElement",
  "HTMLMapElement": "HTMLElement",
  "HTMLMediaElement": "HTMLElement",
  "HTMLMetaElement": "HTMLElement",
  "HTMLMeterElement": "HTMLElement",
  "HTMLModElement": "HTMLElement",
  "HTMLObjectElement": "HTMLElement",
  "HTMLOListElement": "HTMLElement",
  "HTMLOptGroupElement": "HTMLElement",
  "HTMLOptionElement": "HTMLElement",
  "HTMLOutputElement": "HTMLElement",
  "HTMLParagraphElement": "HTMLElement",
  "HTMLParamElement": "HTMLElement",
  "HTMLPreElement": "HTMLElement",
  "HTMLProgressElement": "HTMLElement",
  "HTMLQuoteElement": "HTMLElement",
  "HTMLScriptElement": "HTMLElement",
  "HTMLSelectElement": "HTMLElement",
  "HTMLSourceElement": "HTMLElement",
  "HTMLSpanElement": "HTMLElement",
  "HTMLStyleElement": "HTMLElement",
  "HTMLTableCaptionElement": "HTMLElement",
  "HTMLTableCellElement": "HTMLElement",
  "HTMLTableColElement": "HTMLElement",
  "HTMLTableDataCellElement": "HTMLElement",
  "HTMLTableElement": "HTMLElement",
  "HTMLTableHeaderCellElement": "HTMLElement",
  "HTMLTableRowElement": "HTMLElement",
  "HTMLTableSectionElement": "HTMLElement",
  "HTMLTemplateElement": "HTMLElement",
  "HTMLTextAreaElement": "HTMLElement",
  "HTMLTimeElement": "HTMLElement",
  "HTMLTitleElement": "HTMLElement",
  "HTMLTrackElement": "HTMLElement",
  "HTMLUListElement": "HTMLElement",
  "HTMLUnknownElement": "HTMLElement",
  "HTMLVideoElement": "HTMLElement",
  "ImageData": null,
  "KeyboardEvent": "Event",
  "Location": null,
  "MessageEvent": "Event",
  "MouseEvent": "Event",
  "NamedNodeMap": null,
  "Navigator": null,
  "Node": null,
  "NodeFilter": null,
  "NodeIterator": null,
  "NodeList": null,
  "Performance": null,
  "PerformanceTiming": null,
  "ProcessingInstruction": null,
  "ProgressEvent": "Event",
  "Range": null,
  "Screen": null,
  "Storage": null,
  "StorageEvent": null,
  "TestBinding": null, // XXX
  "TestBindingProxy": null, // XXX
  "Text": null,
  "TextDecoder": null,
  "TextEncoder": null,
  "TreeWalker": null,
  "UIEvent": "Event",
  "URLSearchParams": null,
  "ValidityState": null,
  "WebGLRenderingContext": null,
  "WebGLUniformLocation": null,
  "WebGLBuffer": "WebGLObject",
  "WebGLFramebuffer": "WebGLObject",
  "WebGLRenderbuffer": "WebGLObject",
  "WebGLTexture": "WebGLObject",
  "WebGLProgram": "WebGLObject",
  "WebGLShader": "WebGLObject",
  "WebGLObject": null,
  "WebGLActiveInfo": null,
  "WebGLShaderPrecisionFormat": null,
  "WebSocket": null,
  "Window": null,
  "Worker": null,
  "WorkerGlobalScope": null, // #2823
  "WorkerLocation": null, // #2823
  "WorkerNavigator": null, // #2823
  "XMLHttpRequest": null,
  "XMLHttpRequestEventTarget": null,
  "XMLHttpRequestUpload": null,
};

function createInterfaceMap() {
  var interfaceMap = {};

  function addInterfaces(interfaces)
  {
    for (var entry of interfaces) {
      interfaceMap[entry] = true;
    }
  }

  addInterfaces(ecmaGlobals);
  addInterfaces(Object.keys(interfaceNamesInGlobalScope));
  addInterfaces(["EventWatcher"]);
  return interfaceMap;
}

test(function() {
  var interfaceMap = createInterfaceMap();
  for (var name of Object.getOwnPropertyNames(window)) {
    if (!/^[A-Z]/.test(name)) {
      continue;
    }
    assert_true(name in interfaceMap,
                "If this is failing: DANGER, are you sure you want to expose the new " +
                "interface " + name + " to all webpages as a property on the window? " +
                "Do not make a change to this file without review from jdm or Ms2ger " +
                "for that specific change!");
    if (name in interfaceMap) {
      if ( name in interfaceNamesInGlobalScope ) {
        if ( interfaceNamesInGlobalScope[name] === null ) {
          assert_true(window[name].__proto__.name === "",
              "Interface " + name + " should not inherit from any interface, " +
              " found " + window[name].__proto__.name);
        } else {
          assert_true(window[name].prototype instanceof window[interfaceNamesInGlobalScope[name]],
              "Interface " + name + " should inherit from " + interfaceNamesInGlobalScope[name] + ", " +
              " found " + window[name].__proto__.name);
        }
      }
      delete interfaceMap[name];
    }
  }
  for (var name of Object.keys(interfaceMap)) {
    assert_true(name in window, name + " should be defined on the global scope");
  }
  assert_equals(Object.keys(interfaceMap).length, 0,
                "The following interface(s) are not enumerated: " +
                Object.keys(interfaceMap).join(", "));
});
</script>
